lambda/ingestion/requirements.txt
psycopg2-binary==2.9.7
pandas==2.0.3
boto3==1.28.57
openpyxl==3.1.2
